requirements.txt
dj-database-url==0.5.0
Django==1.11
djangorestframework==3.7.0
gunicorn==20.0.4
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
whitenoise==3.3.1
